Reorganizes components
